refactor(auth): ScuAuthService 重命名为 ScuApiService
- 类名 ScuAuthService → ScuApiService（更准确：认证+数据请求）
- 文件夹 scu_auth/ → scu_api/
- 5 个 part 文件重命名
- barrel export 路径 scu_auth_service.dart → scu_api_service.dart
- extension ScuAuthSchedule/Grades/Classroom → ScuApiSchedule/Grades/Classroom
- 13 个文件的 import 路径更新
- ScuAuthSession / ScuAuthProvider 类名保持不变
- 文档 docs/decisions/auth-architecture.md 同步更新
