+++ lib/pages/course/course_schedule_setting.dart
@@ -6,7 +6,7 @@ import 'package:bugaoshan/models/course.dart';
 import 'package:bugaoshan/pages/course/time_slot_setting_page.dart';
 import 'package:bugaoshan/providers/course_provider.dart';
 import 'package:bugaoshan/providers/scu_auth_provider.dart';
-import 'package:bugaoshan/services/scu_auth_service.dart';
+import 'package:bugaoshan/services/scu_api_service.dart';
 import 'package:bugaoshan/widgets/common/styled_card.dart';
 
 class CourseScheduleSetting extends StatefulWidget {
