+++ lib/providers/scu_auth_provider.dart
@@ -9,7 +9,7 @@ import 'package:bugaoshan/providers/profile_labels_provider.dart';
 import 'package:bugaoshan/providers/secure_storage_provider.dart';
 import 'package:bugaoshan/services/auth/auth_manager.dart';
 import 'package:bugaoshan/services/ocr_service.dart';
-import 'package:bugaoshan/services/scu_auth_service.dart';
+import 'package:bugaoshan/services/scu_api_service.dart';
 
 const _keyAutoLogin = 'scu_auto_login';
 const _keyUserRealname = 'scu_user_realname';
@@ -49,7 +49,7 @@ class ScuAuthProvider extends ChangeNotifier {
   bool get isAutoLoggingIn => _isAutoLoggingIn;
   bool get isLoggedIn => _authManager.isScuLoggedIn;
   bool get isExpired => _authManager.scu.isExpired;
-  ScuAuthService get service => _authManager.scu.service;
+  ScuApiService get service => _authManager.scu.service;
 
   Future<void> login({
     required String username,
