+++ lib/services/scu_api_service.dart
@@ -1,0 +1,4 @@
+// Barrel export — 保持旧 import 路径兼容。
+export 'package:bugaoshan/services/scu_api/scu_api_service.dart';
+export 'package:bugaoshan/services/scu_api/scu_api_models.dart';
+export 'package:bugaoshan/services/scu_api/cookie_client.dart';
