--- lib/services/scu_auth_service.dart
@@ -1,4 +1,0 @@
-// Barrel export — 保持旧 import 路径兼容。
-export 'package:bugaoshan/services/scu_auth/scu_auth_service.dart';
-export 'package:bugaoshan/services/scu_auth/scu_auth_models.dart';
-export 'package:bugaoshan/services/scu_auth/cookie_client.dart';
